Extract arduino connection logic to presenter
